Setup absolute paths
