--- src/App.tsx
@@ -1,9 +1,0 @@
-const App = () => {
-  return (
-    <div className="App">
-      <p>it is working</p>
-    </div>
-  );
-}
-
-export default App;
