+++ src/components/App/index.tsx
@@ -1,0 +1,9 @@
+const App = () => {
+    return (
+        <div className="App">
+            <p>it is working</p>
+        </div>
+    );
+};
+
+export default App;
